First draft of a source predicate
This is a first draft of a source predicate, that enables routing
decisions based on the source of a request.

Right now it is build (and registered) as a custom predicate. We could
discuss adding it to the built in predicates or leaving it like that as
a nice and simple example of how to add custom predicates to skipper.

Also, the test cases might not be complete. I'm open to suggestions for
more test cases/corner cases, that are not yet tested.