Basic per backend host metrics
The background of this PR is the following:

We would like to create a dashboard, that shows us a distribution of where
skipper routes the requests to. Skipper currently supports metrics on a
per-backend base, but only keyed via the route-id. For our use-case, we would
need a mapping from route-id to backend-host. This is very brittle.

This is a simple solution, that would also provide a metric based on the
hostname of the backend. This would allow us to aggregate in ZMON. We could for
example create three buckets:
1. all requests, that go to *.example.org
2. all requests, that go to *.example.com
3. all others.
Plotting the request count over those three buckets would be exactly what
we want.

I tried to keep this as general as possible to be useful to others as well and
made it opt-in, because it might not be interesting for all users of skipper.