Swarm ratelimiter (#593)
* communication between nodes using swim protocol via Hashicorp's memberlist
* current swarm types: Kubernetes and Fake (testing purposes), these types are needed to find your initial peers.
* implemented  use cases: `clusterRatelimit` and `clusterClientRatelimit`, that enable you to specify number of requests per `time.Duration` for a swarm of skipper instances, that can grow and shrink without interfering much with the ratelimit setting. 

Why we do not use a central ratelimit service? Because you have to increment and query this service, for every request, that means you created the biggest central service that you rely on. This implementation shares the current data for every request and calculates locally the current state, which asynchronously gets updated in the not hot path of the request handling.  

Signed-off-by: Sandor Szücs <sandor.szuecs@zalando.de>