Fix the Host predicate regex example
The docs state that regular expressions should be surrounded by slashes
(`/`), but just below, the `Host` example is surrounded by double
quotes.

In the `Host` paragraph, there are both slashes and quotes; this example
will fail. I have changed them to just slashes, for consistency with the
description at the top of the page.

However, I will not touch the `PathRegexp` examples, as I have not
tested them yet. Moreover, having a slash-intense literal surrounded by
slashes might not be the clearest syntax.