Set URL from Host and TLS in dynamic backend
When using dynamic backend, it is expected to use one of the
`setDynamicBackend*` filters to fill the host and scheme for the request
URL. However, there are cases in which we don't want to set it explicitly,
for example:

    foo: * -> compress("text/html") -> <dynamic>

Now if I do something like:

    curl -x http://skipper:9090 http://httpbin.org/anything

it complains about URL scheme and host not being set.

In this commit, the host and scheme part are automatically set for the
request URL before they're set from the statebag. This still allows the
user to modify host and scheme using the `setDynamicBackend*` filters
but if let unset, it will fallback to some sane defaults.

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>