Change OpenTracing parameter structure to accommodate more options
Currently, skipper creates spans that serializes to ~2.5 KB of payload in
protobuf format based on my experiments. The main contributors to this are
`skipper.route` tag, logs for marking filter start & done events, and the
response stream events.

We can reduce the size by dropping these attributes however we may lose
debugging abilities. To get around this, I propose we add feature flags under
`OpenTracingParams` that would be passed to `Proxy`. In this way, we can
enable logging these attributes when it is needed and keep it turned off by
default.

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>