Add options to reduce the payload shipped per request (#1066)
* Change OpenTracing parameter structure to accommodate more options

Currently, skipper creates spans that serializes to ~2.5 KB of payload in
protobuf format based on my experiments. The main contributors to this are
`skipper.route` tag, logs for marking filter start & done events, and the
response stream events.

We can reduce the size by dropping these attributes however we may lose
debugging abilities. To get around this, I propose we add feature flags under
`OpenTracingParams` that would be passed to `Proxy`. In this way, we can
enable logging these attributes when it is needed and keep it turned off by
default.

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Add option to log request filter lifecycle events or not for proxy

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Log response filter lifecycle events as well

This will be only logged in case `OpenTracingParams.LogFilterEvents` flag is set

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Unify how tags are set

and use DefaultIncludedTags for denoting the tags enabled by default

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Invert the logic to include tags to denote what should be excluded

I think we will want to exclude tags rather than include everything but specific
ones

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Pass opentracing related options for excluding tags & filter events

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Extend tests to include new functionality to enable/disable tags & logs

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Drop commented out code

Default tracer for `Proxy` is now handled in `proxy/tracing.go`.

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Add comment on OpenTracing related options

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Add flag to control events logged for the HTTP response stream

These events mark the times response headers & chunks of the payload is sent to
the client

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>

* Simplify how excluded tags are handled

Signed-off-by: Serbay Arslanhan <serbay.arslanhan@zalando.de>