Fixed issues reported by the maintainers
Most notable change is that the http.Transport is no longer copied when
proxy is used. Instead, the common transport is initialized with a
proxy function that checks for a particular header in the request
`X-Skipper-Proxy` which holds the proxy URL that should be used.

Although it is possible to easily clone the http.Transport without
touching the mutexes, I decided against it because if a new transport
was created for each request it would lose some benefits like connection
caching that help with the performance.

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>