Added new filter backendIsProxy (#1252)
* Added new filter backendIsProxy

The filter notifies the proxy that the backend this request is going to be sent to is also
proxy. This means that the request is going to be sent the usual proxy way:

    // HTTP request
    GET http://example.com HTTP/1.1
    ...

    // HTTPS request
    CONNECT https://example.com HTTP/1.1
    ...

Example:

    foo1:
      *
      -> backendIsProxy()
      -> "http://proxy.example.com";

    foo2:
      *
      -> backendIsProxy()
      -> <roundRobin, "http://proxy1.example.com", "http://proxy2.example.com">;

    foo3:
      *
      -> setDynamicBackendUrl("http://proxy.example.com")
      -> backendIsProxy()
      -> <dynamic>;

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>

* Fixed issues reported by the maintainers

Most notable change is that the http.Transport is no longer copied when
proxy is used. Instead, the common transport is initialized with a
proxy function that checks for a particular header in the request
`X-Skipper-Proxy` which holds the proxy URL that should be used.

Although it is possible to easily clone the http.Transport without
touching the mutexes, I decided against it because if a new transport
was created for each request it would lose some benefits like connection
caching that help with the performance.

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>

* Fixed failing test

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>

* Default to incoming request Host

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>

* Fixed tests

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>

* Minor fixes requested by maintainers

Signed-off-by: Stepan Bujnak <stepan.bujnak@gmail.com>