Explain when using predicate as filter
For beginners, it's very easy to mix predicates and filters. Also when using Kubernetes annotations, sometimes a mistake is easy to make.

Instead of returning `failed to process route (kube_x): filter not found: 'PathRegexp'` it will now return `failed to process route (kube_x): trying to use 'PathRegexp' as filter, but it is only available as predicate`

Signed-off-by: Ruud Kamphuis <ruudk@users.noreply.github.com>