Log proxy errors in access log
Logging the proxy errors without the request context is pretty useless. The only way to find out which request failed is to enable flowId's and try to map them together. On a production server with lots of requests this is almost impossible.

To make it easier to find out which request failed, the proxy errros are now stored in the Context.

When the access log is enabled, the proxy error is appended to the end of the access log.

Signed-off-by: Ruud Kamphuis <ruudk@users.noreply.github.com>