flexible upstream headers
Similar to [Keycloak Gatekeeper Custom Claim Headers](https://github.com/keycloak/keycloak-documentation/blob/master/securing_apps/topics/oidc/keycloak-gatekeeper.adoc#custom-claim-headers) we need Skipper to inject headers to upstream endpoints to provide user identity.

As of now the "Skipper-Oidc-Info" header provides this information but in a custom, not ready to use format which would require to adapt each backend application.

Like Gatekeeper this information should be possible to be specified per use case on the ingress annotation level.

This implementation uses the `GJSON` syntax already used for other purposes to define and parse the information that is requested and inject as the defined header name.

For backward compatibilty, this additional field is optional, falling back to outputting the "Skipper-Oidc-Info" as before.

Signed-off-by: Samuel Lang <gh@lang-sam.de>