Deduplicate OIDC cookies
For each ingress Skipper will hash a new cookie. If several ingresses for the same domain are defined,
either by several hosts or sub-paths, it will accumulate cookies on the client browser which can lead to side effects like header limits exceeded or heavy request payloads.

There might be better approaches to tackle this issue by specifying the actual path of its ingress on the cookie,
however this information is apparently not available at the filter level for now.

This simple approach removes the `RedirectURL` out of the hashing, which keeps the idP, client-ID, client-secret, scopes,
claims and optional parameters as a sufficient source of integrity signature.
It will allow Skipper to reuse the cookie where applicable and deduplicate where avoidable.

Signed-off-by: Samuel Lang <gh@lang-sam.de>