Fully functional OAuth2 grant flow.
Fixes and extends the original grant PR for a fully functional OAuth2
authorization code grant flow.

Add support for loading and automatic reloading of client credentials
from the filesystem. This is accomplished by using the preexisting
`SecretsProvider`.

Make the feature more configurable. For example:
* provide arbitrary key-value pairs which should be set for all calls
to authorize and token endpoints. Accomplished with the help of a new
`mapFlags` option type.
* name of the header where to place the user's access token.
* name of the cookie where to store encrypted tokens.

Create a new `grantClaimsQuery()` filter. This is an alias for
`oidcClaimsQuery()` and enables the combination of access control rules
with the `oauthGrant()` filter.

Increase OAuth grant flow state validity time from 1 minute -> 1 hour.
One minute is a short time. The user's login is disrupted and must be
restarted if it takes them longer than a minute to actually log in.
By increasing this to 1 hour the user experience improves without
undermining security much.

Fix invalid redirect URI given to the token callback.

Fix broken token refresh. When token is valid, but token refreshing
fails, the code crashes with a segfault. The fix involves not setting
the state bag oauth2 token to a nil value.

Fix cookie expiration. Expiration relies on access token expiry,
meaning cookie is deleted by the browser before refresh token expires.
This is fixed by adding an extra 30 days to the cookie expiration time.

Reuse the `TokeninfoClient` for fetching token claims and remove the
duplicated code.

Refactor token validity checking and token refreshing code for improved
readability.

Add the tokeninfo payload to the state bag under the `tokeninfoCacheKey`
so the `oauthGrant()` filter can be used with the `forwardToken()`
filter.

Add support for handling multiple cookies with the same name. It's
possible that the browser calls `oauthGrant()` with multiple cookies
under the same name as the grant token cookie. The code handles it by
attempting to decode each cookie with the correct name, until it finds
one that decodes properly.

Make sure the grant token cookie is dropped from all requests that it
forwards downstream to prevent the cookie from leaking.

Add tutorial documentation for using the OAuth2 auth code grant flow.

Add reference documentation for all grant-related filters.

Add unit tests for new code and unit test the existing token
refresh functionality.

Signed-off-by: Olli Salonen <oasalonen@gmail.com>