Add support for revoking grant tokens.
The new `grantLogout()` filter allows defining routes that call the
OAuth2 provider's token revocation endpoint. It attempts to revoke both
the access and refresh tokens.

According to RFC 7009, the provider does not need to support access
token revocation. The filter supports this case by ignoring a 401 error
with a `unsupported_token_type` error when revoking the access token.

The filter also instructs the client to delete the grant token cookie
by setting a `Set-Cookie` header with an empty value and a `MaxAge` of
`-1`.

In order to use the filter, Skipper must be provided the token
revocation endpoint URL using `-oauth2-revoke-token-url`.

Signed-off-by: Knut Hühne <knut.huehne.external@zalando.de>