Do not export createCookie from auth
The function to create the cookie was only exported because
it was used in the tests that live in the auth_test package.

The function is now no longer exported. For the tests to still
be able to get test cookies, `grantcookie_test.go` was introduced
which contains three public helper functions that can be used
in the tests where different kinds of cookies are required.

Signed-off-by: Knut Hühne <knut.huehne.external@zalando.de>