Match ingress to service using endpoints
This commit changes matching of ingresses to services to use the
containerPort intead of the targetPort. Using the targetPort can lead to
skipper being unable to route traffic to the target.

Two tests in TestConvertPathRule were failing but are not valid cases to
test as Kubernetes will not allow you to create a service with two ports
of the same name. They have been removed.

Signed-off-by: Markus Cisler <hello@markuscisler.com>