Match ingress to service using endpoints (#1646)
* Match ingress to service using endpoints

This commit changes matching of ingresses to services to use the
containerPort intead of the targetPort. Using the targetPort can lead to
skipper being unable to route traffic to the target.

Two tests in TestConvertPathRule were failing but are not valid cases to
test as Kubernetes will not allow you to create a service with two ports
of the same name. They have been removed.

Signed-off-by: Markus Cisler <hello@markuscisler.com>

* Add test for matching ports on services

Signed-off-by: Markus Cisler <hello@markuscisler.com>

* Remove getServicePort tests

They are replaced by the fixtures in
dataclients/kubernetes/testdata/ingress/service-ports.

Signed-off-by: Markus Cisler <hello@markuscisler.com>

* Fix endpoint values in tests

Signed-off-by: Markus Cisler <hello@markuscisler.com>

* If only one port exists in endpoint, use it

Otherwise match port by name

Signed-off-by: Markus Cisler <hello@markuscisler.com>

* Match container spec port number in test

Signed-off-by: Markus Cisler <hello@markuscisler.com>