Remove unnecessary loop on k8s internal hosts (#1758)
In #1676 a new loop for the internal hosts were introduced by a first
implementation and not removed. The initial implementation assumed that
multiple domains with different CIDRs could be configured, therefore,
required different route for every host. The implementation changed
during the PR discussion and with the current one multiple domains can
be set but the predicates are all added to the routes without
distinction of domain and the loop is unnecessary. Check [this
comment][0] for more details.

This commit removes the unnecessary loop and create the routes combining
all internal hosts.

[0]: https://github.com/zalando/skipper/pull/1676#discussion_r614897845

Signed-off-by: Jonathan Juares Beber <jonathanbeber@gmail.com>