Set OIDC cookie maxAge according to claims exp
Currently, Skipper hard-codes the maxAge of the OIDC cookie to one
hour, even if the expiration time of the claim is longer than that.
This results in early re-authentication, even though it might not really
be necessary yet.

With this change, the exp field is attempted to be parsed.
If it's successful, it will use that value as the maxAge for
the OIDC cookie.

Signed-off-by: Andreas Skorczyk <me@andreas-sk.de>