Clarify naming of test fixtures, remove duplicates
It was not clear why fixtures were prefixed with "single" and "multiple"
in the first place. Both test service port mappings.

The fixtures prefixed with "multiple" were removed as they did not
improve coverage at all. Overall the fixtures were reduced to
"named.yaml" and "no-name.yaml" to test correct behaviour with named and
unnamed ports.

Relates to #1902.

Signed-off-by: Markus Cisler <hello@markuscisler.com>