scheduler: ensure single lifo instance per route
Multiple `lifo` filter instances on the route share the same job queue but claim multiple slots per request.
This is not obvious when users configure `lifo` filter in the ingress manifest and another instance is added
by the `-default-filters-prepend` flag.

The change keeps only the last `lifo` filter instance on the route to enable default configuration override.

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>