scheduler: delays closing of deleted lifo queues
Closing of a lifo queue on route deletion while in-flight requests try
to get a slot from it results in `Unknown error for route based LIFO: queue closed for host`
error and 500 response to the client.

To mitigate the problem this change puts deleted queues aside and closes them after one minute.
The proper fix would be to close queue when all route's in-flight requests complete, see #202.

Updates #1851

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>