clear invalid cookies
If OIDC cookies are invalid, the client is redirected but the cookies are not cleared.
In case that the new cookie will produce smaller set of chunks, an invalid chunk is not overwritten and kept in the browser and Skipper will try to attach this junk to the new set.

This changes the behaviour in that regard that it clears cookies with relevant names in case of a new auth session.

Signed-off-by: Samuel Lang <gh@lang-sam.de>