Fill the authuser field of the access log (#2143)
* Fill the authuser field of the access log

The auth filter already stores the username of the identified user in
the state bag. This PR forwards this value to the access log filter by
storing it in the AccessEntry structure.

It is then used in the `authuser' field of the Common Log Format
formatter. It can also be accessed by other formatters (such as the
JSON formatter) via the `auth-user' field of the logrus Entry.

Signed-off-by: Adrien Surée <adrien.suree@zalando.de>