Remove sentence stating the necessity of go 1.11
What?

Remove the sentence describing that the used needed go version >= 1.11
because of go modules.

Why?

This sentence was misleading since currently you need go version >= 1.19
and this was added just to make clear the necessity of go modules which
by now is something that can be assumed, since the go project use go
modules by default for quite some time now.

The idea to just remove it instead to update the version is that this is
something pretty straight forward to find out and without this sentence
here we don't need to remember to update the docs everytime we update
the version.

-----

Signed-off-by: Lucas Thiesen <lucas.thiesen@zalando.de>