dataclients/kubernetes: do not create catchall routes
Catchall routes are created for Ingresses and RouteGroups that
do not define predicates matching all requests for the host.

Initially catchall routes were introduced by #425 for #408.
Later it was discovered that multiple Ingresses for the same host
may create conflicting catchall routes which was fixed by #436.

Addition of RouteGroups #1180 introduced catchall routes for them as well
but did not take into account that and Ingress and RouteGroup
may use the same host which results in a catchall route created
e.g. for RouetGroup that interferes with routes created for Ingress.

This change removes catchall route logic because it simply shunts the request
and responds with the default response status which matches the behaviour
when Skipper can not look up route at all.

Updates #2394

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>