minor: countPathRoutes function
The name could be a little bit misleading because actually this function
calculates the number of Path{,Subtree} predicates in the single route.
Nobody calculates the number of routes here, there is only one route
which this function takes as input.
Therefore n variable corresponds to the number of predicates, not the
routes how it seems to be when taking a look at the code.

Signed-off-by: Roman Zavodskikh <roman.zavodskikh@zalando.de>