Fix list formatting in filters.md (#3188)
The Markdown processor used on the reference docs (e.g. https://opensource.zalando.com/skipper/reference/filters/#awssigv4) requires an empty line before a list to recognize it as such (and not have the content part of the paragraph before it). This adds two such line breaks.

(The difference is not visible on the Github preview, as its processor doesn't require the empty line.)

Signed-off-by: Paul Ebermann <paul@zalando.de>