proxy: trace response body size and request header length
Add response body size and request headers length rounded to the next power of two as proxy span tags.

The rounding is done to reduce cardinality of the tags and avoid
introducing new configuration parameter.

Since it is tricky to measure request header length in bytes it is
calculated as the sum of lengths of header names and values.

Note that OpenTelemetry [Semantic Conventions for HTTP Spans](https://opentelemetry.io/docs/specs/semconv/http/http-spans/)
define the following related attributes:
* experimental [http.response.body.size](https://opentelemetry.io/docs/specs/semconv/attributes-registry/http/#http-response-body-size) for body size in bytes
* stable [http.request.header.<key>](https://opentelemetry.io/docs/specs/semconv/attributes-registry/http/#http-request-header) for header value

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>