Add acccess log formatter (#3458)
This PR adds the ability to use a custom access log formatter.

Currently, accessLogs can be configured in three ways:
- If `AccessLogJSONEnabled` is defined, you can either pass a JSONFormatter, or have Skipper initialize one.
- If `AccessLogJSONEnabled` is not defined, it uses a basic text formatter.

There are cases (for example, when configuring NewRelic logrus integration, [see here](https://github.com/newrelic/go-agent/tree/master/v3/integrations/logcontext-v2/nrlogrus), that a custom formatter is needed, but current logger can't be configured.

My initial approach was to change type on `AccessLogJsonFormatter` (typed parameter in Skipper options) to be able to pass any logrus.Formatter (interface), as it wouldn't break anything. But it would be weird to have something called `JSONFormatter` that may not be a JSON formatter.

Signed-off-by: Xavi Ivars <xavi.ivars@gmail.com>