filters/accesslog: Fix type assertion in AccessLogFilter.Request
The commit corrects a bug in the AccessLogFilter.Request method where
an incorrect type assertion was being performed. The code was
attempting to assert a value from the state bag as an AccessLogFilter
value rather than a pointer (*AccessLogFilter), which would cause
runtime errors when two or more instances of the filter are used.

Signed-off-by: Aleksandr Ponimaskin <aleksandr.ponimaskin@zalando.de>