feat/metric: add new metric for skipper latency
This change involves adding a new mtric to measure the overall skipper latency
of a request. i.e., the time taken by a request in skipper itself, so this does not
include time taken for serving for the response to client and the time taken for the backend
round trip.

Signed-off-by: speruri <surya.srikar.peruri@zalando.de>