Allow configuration of bundle download behaviour of open-policy-agent filter (#3464)
Addresses: https://github.com/zalando/skipper/issues/3119

**Solution**
add option to override the period trigger for discovery and bundle
plugins
- control timing of f.ex. bundle download directly (also adding jitter
to avoid different opa instances updating at the same time)
- fail fast for non-recoverable errors

**Context**
The current (and future default) behaviour is relying on the
configurable startup timeout to determine if the open-policy-agent
filter can successfully serve requests (if it was able to download the
bundles, aso).
The solution is based on a [change](add option to override period
trigger for plugin) in upstream opa that exposes the httpError struct on
the trigger method.

---------

Signed-off-by: Torsten Wunderlich <torsten.wunderlich@zalando.de>