routing: ignore duplicate dataclients
Ignore duplicate dataclients since this is likely a misconfiguration.

Note that datasource still allows multiple pointer dataclients of the same type
or non-equal value-type dataclients.

Routing constructor does not return error and we generally do not use panic
therefore ignoring and logging an error is the only simple option.

This is also needed because datasource manages routes keyed by DataClient and
to correctly signal first route load.

Follow up on #3447

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>