feat/metrics : add new metrics for measuring skipper latency (#3506)
This change involves adding a new metric to measure the overall skipper
latency of a request. i.e., the time taken by a request in skipper
itself, so this does not include time taken for serving for the response
to client and the time taken for the backend round trip.

skipper latency = serve duration - response duration - backend duration

---------

Signed-off-by: speruri <surya.srikar.peruri@zalando.de>
Co-authored-by: speruri <surya.srikar.peruri@zalando.de>