Added check to handle token parsing with claims without "sub". (#3287)
### Description 
In the oidc_introspection.py `func (filter *oidcIntrospectionFilter)
Request(ctx filters.FilterContext)` function get value form the Claim
map using `sub` key and then casts the value into a string, but this
would panic if no sub key is present. I added a check to ensure we only
cast values if sub key is present.

### Linked Issue 
closes #3216 

### Approach to solution
As describe in the following
[comment](https://github.com/zalando/skipper/issues/3216#issuecomment-2433498646),
I dug a bit deeper and found 2 suitable values to use if subject is not
present in token. One is the Subject value in tokenContainer struct or
UserInfo Struct. I decided to go with the tokenContainer struct. If this
is not correct please do guide me if UserInfo Struct would be better.

---------

Signed-off-by: wassafshahzad <wassafshahzad@gmail.com>