proxy/tracing: loopback spans bugfix (#3544)
When a request is looped back, all tracing operations (such as tagging
and creating new spans) are performed on the original request's span
rather than on the loopback span. This results in inaccurate tracing
data: tags are overwritten, and the ingress span incorrectly appears to
have child spans that actually belong to the inner loopback spans.

This PR adds a test case to illustrate the issue, along with a bug fix.


Signed-off-by: Aleksandr Ponimaskin <aleksandr.ponimaskin@zalando.de>