fix: Websocket requests fail if case of value in Connection header and case of key in Upgrade header do not match (#3563)
When a client sends a request to upgrade to a Websocket connection
and presents the value of the header `Connection` in lower-case
then Skipper fails to upgrade the request.

This change fixes the issue by using a function to read the header that
is case insensitive.

The bug was likely introduced in #3553.

Signed-off-by: Markus Meyer <m.meyer@mytaxi.com>