fix: jwt predicates with regexp cached on create.
Routing table updates are frequently in skipper. We assume every 3s a rebuild of the routing tree.
Therefore a leak in objects like predicate creation can visibly leak memory.
As written in https://github.com/zalando/skipper/issues/3609 and mentioned in https://github.com/zalando/skipper/pull/3600/files\#r2313466897 we appended unbounded and never cleand up predicates in the registry
Now we use a map to also cache creation of the predicate to not grow unbounded and this also speedsup predicate creation and reduces number of objects, so allocations and memory if the same eskip.Predicate string would be used often.

Signed-off-by: Sandor Szücs <sandor.szuecs@zalando.de>