fix: threaded notifications (#3704)
This commit fixes the reference to the output of the step. 
Also moving the check-race step further in pipeline, it is cheaper to
fail earlier in case if vulnerability detected.
The query argument moved to the secret itself.

<img width="399" height="388" alt="Screenshot 2025-11-07 at 13 38 52"
src="https://github.com/user-attachments/assets/eac34cd1-609a-46dc-9716-5e1b3d1c41f0"
/>

Signed-off-by: Aleksandr Ponimaskin <aleksandr.ponimaskin@zalando.de>