fix: `TestPHCNoHealthyEndpoints` use roundRobin for load balancing (#3716)
`TestPHCNoHealthyEndpoints` uses `random` for load balancing, and it's
possible some endpoints are never visited. A never visited endpoint
would have a drop probability of 0 (i.e., the endpoint is allowed to
pass the request). If we use `roundRobin`, we remove the uncertainty in
request distribution and we have a better exepectation on the metrics
count.

The detailed reasoning can also be found in:
https://github.com/zalando/skipper/issues/3189#issuecomment-3509302430

fix #3189

Signed-off-by: Max Qian <zq2@illinois.edu>