Fix: flakytest admission control (#3725)
The core problem was that we have had a non-seeded global random source
that generated every time different random numbers so it's easy to have
bad luck. Another issue was that if you have a global random source is
that the different test cases interfere with each other. The third
problem was that we used the same global random to have randomized
response codes on the server side in the httptest.Server.
All of it is now fixed. 

We have an option to use a static seeded random source by filter that we
use for tests.
In general we use the global random source that we have to guard by a
mutex at least `test -race` says so.
Maybe we should also create filter local random source for the
production build but I think it's fine like this.

Signed-off-by: Sandor Szücs <sandor.szuecs@zalando.de>