metrics: remove invalid routes deletion (#3776)
When validation is enabled and a new invalid route is created, it gets
rejected and a metric is emitted. But when the route is later fixed, the
old metric is not removed. Once the route is fixed, there is no error,
so the corrected route no longer matches the old metric key.

Signed-off-by: Veronika Volokitina <v.volokitinaa@gmail.com>